portfolio-chatbot/requirements.txt
Flask==3.0.3
google-generativeai>=0.7.0
python-dotenv>=1.0.1
gunicorn==21.2.0